login-auth-system/requirements.txt
flask
mysql-connector-python
flask-session